+++ src/main/java/com/api/hotifi/session/web/response/AmountToBePaidResponse.java
@@ -4,11 +4,13 @@ import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @AllArgsConstructor
 public class AmountToBePaidResponse {
 
-    private double amountToBePaid;
+    private BigDecimal amountToBePaid;
 
 }
